requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pydantic
python-dotenv
python-multipart
pandas
openai
motor
python-jose[cryptography]
passlib[bcrypt]